rename `StatBoxList` -> `StatCards`
