+++ packages/app/src/library/StatCards/Timeleft.tsx
