+++ packages/app/src/pages/Nominate/Active/Stats/MinimumActiveStake.tsx
@@ -3,7 +3,7 @@
 
 import { useApi } from 'contexts/Api'
 import { useNetwork } from 'contexts/Network'
-import { Number } from 'library/StatBoxList/Number'
+import { Number } from 'library/StatCards/Number'
 import { useTranslation } from 'react-i18next'
 import { planckToUnitBn } from 'utils'
