+++ packages/app/src/pages/Nominate/NominationGeo/Stats/AnalyzedPayouts.tsx
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import { useNetwork } from 'contexts/Network'
-import { Number } from 'library/StatBoxList/Number'
+import { Number } from 'library/StatCards/Number'
 import { useTranslation } from 'react-i18next'
 
 import type { AnalyzedPayoutsProps } from '../types'
