+++ packages/app/src/pages/Overview/Stats/AveragelRewardRate.tsx
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import { useAverageRewardRate } from 'hooks/useAverageRewardRate'
-import { Text } from 'library/StatBoxList/Text'
+import { Text } from 'library/StatCards/Text'
 import { useTranslation } from 'react-i18next'
 
 export const AverageRewardRate = () => {
