+++ packages/app/src/pages/Overview/Stats/NextRewards.tsx
@@ -8,7 +8,7 @@ import { useApi } from 'contexts/Api'
 import { useNetwork } from 'contexts/Network'
 import { fromUnixTime, getUnixTime } from 'date-fns'
 import { useEraTimeLeft } from 'hooks/useEraTimeLeft'
-import { Timeleft } from 'library/StatBoxList/Timeleft'
+import { Timeleft } from 'library/StatCards/Timeleft'
 import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import { formatTimeleft } from 'utils'
