+++ packages/app/src/pages/Validators/Stats/ActiveValidators.tsx
@@ -4,7 +4,7 @@
 import BigNumber from 'bignumber.js'
 import { useApi } from 'contexts/Api'
 import { useStaking } from 'contexts/Staking'
-import { Pie } from 'library/StatBoxList/Pie'
+import { Pie } from 'library/StatCards/Pie'
 import { useTranslation } from 'react-i18next'
 import { percentageOf } from 'ui-graphs'
