+++ packages/app/src/pages/Validators/Stats/TotalValidators.tsx
@@ -3,7 +3,7 @@
 
 import BigNumber from 'bignumber.js'
 import { useApi } from 'contexts/Api'
-import { Pie } from 'library/StatBoxList/Pie'
+import { Pie } from 'library/StatCards/Pie'
 import { useTranslation } from 'react-i18next'
 import { percentageOf } from 'ui-graphs'
